[INFER] Add dependency on type of pushed value in JSOP_GETGNAME, bug 649261, mark properties redefined with getters/setters as configured, bug 649272.
UltraBlame original commit: a4131835b866364816d60729d80f27c294180b1f

+++ js/src/methodjit/FastBuiltins.cpp
@@ -349,8 +349,8 @@ mjit::Compiler::inlineNativeFunction(uint32 argc, bool callingNew)
                            : JSVAL_TYPE_UNKNOWN;
 
     
-    for (int i=0; i<argc; i++) {
-        FrameEntry * arg = frame.peek(-(i+1));
+    for (unsigned i=0; i<argc; i++) {
+        FrameEntry * arg = frame.peek(-((int)i+1));
 
         if (!arg->isTypeKnown())
             return Compile_InlineAbort;
